Add files via upload
This project utilizes the k-Nearest Neighbors (k-NN) algorithm to analyze a cancer dataset and predict the likelihood of cancer based on patient data.
Key Hiughlights:
.)Data Preparation
.)Model Implementation
.)Evaluation Matrix
Results:
The project demonstrates the effectiveness of k-NN in cancer prediction, showcasing accurate results and reliable performance metrics. The confusion matrix aids in assessing the model's strengths and areas for improvement, ensuring meaningful and actionable insights.